% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/alphaAlgoritm.R
\name{alphaAlgorithm}
\alias{alphaAlgorithm}
\title{alphaAlgorithm}
\usage{
alphaAlgorithm(parameters.list = NA)
}
\arguments{
\item{parameters.list}{The list of the parameters used to set the object}
}
\description{
implement alphaAlgorithm
}

