% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/firstOrderMarkovModel.R
\name{firstOrderMarkovModel}
\alias{firstOrderMarkovModel}
\title{firstOrderMarkovModel}
\usage{
firstOrderMarkovModel(parameters.list = list())
}
\arguments{
\item{parameters.list}{the list of the parameter used to set the object}
}
\description{
implement a simple markov Model
}

