% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utils.R
\name{is.included}
\alias{is.included}
\title{is.included.}
\usage{
is.included(a, b)
}
\arguments{
\item{a}{the first set}

\item{b}{the second set}
}
\description{
A very complex function:  is A included in B?
}

