% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappingFunctions.R
\name{LD.builder}
\alias{LD.builder}
\title{Create a \code{dataLoader} object}
\usage{
LD.builder()
}
\value{
a \code{dataLoader} empty object
}
\description{
This function allows to instantiate a \code{dataLoader} object.
}
\examples{
\dontrun{

# Instantiate a 'dataLoader' object
obj<- LD.builder()
}
}

