% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/petriNetworkModel.R
\name{petriNetworkModel}
\alias{petriNetworkModel}
\title{class for parsing, checking and executing Petri Network}
\usage{
petriNetworkModel()
}
\description{
bla bla bla bla
              ri bla bir bla bir bal
}

