% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PManager.R
\name{PManager}
\alias{PManager}
\title{class for handling Process Mining issues}
\usage{
PManager()
}
\description{
Is the class for handling with Process Mining issues.
              Many methods has been developed: please see vignette for more details...
}

